MainUi Gacha Connet

+++ Styx_Station/Assets/03. Scripts/ScriptableObject/Drop/PetDrop.cs
@@ -31,7 +31,7 @@ public class PetDrop : ScriptableObject
         if (sum <= 0)
             return null;
 
-        if (sum > 1)
+        if ((int)sum > 1)
             return null;
 
         var random = Random.Range(0, sum);
